// shared/schema/lease.d.ts
export declare const insertLeaseHeaderSchema: z.ZodObject<{
    id: z.ZodOptional<z.ZodString>;
    status: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    currency: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    termMonths: z.ZodNumber;
    initialDirectCosts: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    prepaidLeasePayments: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    leaseIncentives: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    leaseType: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    assetClass: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    isModified: z.ZodOptional<z.ZodNullable<z.ZodBoolean>>;
    modificationDate: z.ZodOptional<z.ZodNullable<z.ZodDate>>;
    previousLiability: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    modificationReason: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    createdAt: z.ZodOptional<z.ZodNullable<z.ZodDate>>;
    updatedAt: z.ZodOptional<z.ZodNullable<z.ZodDate>>;
} & {
    leaseNumber: z.ZodString;
    description: z.ZodString;
    vendorId: z.ZodString;
    discountRate: z.ZodNumber;
    commencementDate: z.ZodString;
    expirationDate: z.ZodString;
}, "strip", z.ZodTypeAny, {
    description: string;
    expirationDate: string;
    vendorId: string;
    leaseNumber: string;
    termMonths: number;
    commencementDate: string;
    discountRate: number;
    id?: string | undefined;
    createdAt?: Date | null | undefined;
    updatedAt?: Date | null | undefined;
    status?: string | null | undefined;
    currency?: string | null | undefined;
    leaseType?: string | null | undefined;
    initialDirectCosts?: string | null | undefined;
    prepaidLeasePayments?: string | null | undefined;
    leaseIncentives?: string | null | undefined;
    assetClass?: string | null | undefined;
    isModified?: boolean | null | undefined;
    modificationDate?: Date | null | undefined;
    previousLiability?: string | null | undefined;
    modificationReason?: string | null | undefined;
}, {
    description: string;
    expirationDate: string;
    vendorId: string;
    leaseNumber: string;
    termMonths: number;
    commencementDate: string;
    discountRate: number;
    id?: string | undefined;
    createdAt?: Date | null | undefined;
    updatedAt?: Date | null | undefined;
    status?: string | null | undefined;
    currency?: string | null | undefined;
    leaseType?: string | null | undefined;
    initialDirectCosts?: string | null | undefined;
    prepaidLeasePayments?: string | null | undefined;
    leaseIncentives?: string | null | undefined;
    assetClass?: string | null | undefined;
    isModified?: boolean | null | undefined;
    modificationDate?: Date | null | undefined;
    previousLiability?: string | null | undefined;
    modificationReason?: string | null | undefined;
}>;
export declare const insertLeasePaymentSchema: z.ZodObject<{
    id: z.ZodOptional<z.ZodString>;
    leaseId: z.ZodString;
    paymentType: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    frequency: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    description: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    createdAt: z.ZodOptional<z.ZodNullable<z.ZodDate>>;
} & {
    amount: z.ZodNumber;
    startDate: z.ZodString;
    endDate: z.ZodString;
}, "strip", z.ZodTypeAny, {
    amount: number;
    startDate: string;
    endDate: string;
    leaseId: string;
    id?: string | undefined;
    createdAt?: Date | null | undefined;
    description?: string | null | undefined;
    frequency?: string | null | undefined;
    paymentType?: string | null | undefined;
}, {
    amount: number;
    startDate: string;
    endDate: string;
    leaseId: string;
    id?: string | undefined;
    createdAt?: Date | null | undefined;
    description?: string | null | undefined;
    frequency?: string | null | undefined;
    paymentType?: string | null | undefined;
}>;